Upgrade to material-ui 1.2.0, remove and fix features
Remove those features that depended on formsy-material-ui as that
isn't compatible with our stack.
Fix the simple dialog feature to work with material ui v1.
This has an incorrect package.json in the commit. The
2 @gp-technical stack-packs (app and components) should be
present. Currently working with a locally 'yarn link' version
of those components due to problems with react15/react16.